+++ packages/ts-newskit/src/components/slices/shared-styles/index.ts
@@ -121,16 +121,10 @@ export const InlineTextBlock = styled(TextBlock)`
   }
 `;
 
-export const StyledBlock = styled(Block)`
-  position: relative;
-  display: flex;
-  align-items: last baseline;
-  flex-wrap: wrap;
-  ${getSpacingCssFromTheme('row-gap', 'space020')};
-`;
-
-export const Wrapper = styled.div`
-  display: flex;
-  align-items: center;
-  ${getSpacingCssFromTheme('gap', 'space010')};
+export const VideoIconContainer = styled.span`
+  vertical-align: baseline;
+  line-height: 24px;
+  :last-child > div {
+    display: none;
+  }
 `;